Piece Movement "target squares" correctely implemented, with the exception of pawn promotions, castling, En-Passant and awareness of check.
Next steps are to add pawn promotion, castling and En-Passant, and Check awareness. Currently, the allowable moves are based purely on the board position, which will be functional for the promotion moves, and the awareness of check.

To implement the awareness of check, the program must detect whether the next "half move" may allow the opponent to capture the king. This can be identified by storing the ID of every attacked square after a potential move, and checking whether the king shares ID with one of these squares. If so, the move must be eliminated from the possible move pool.

Castling and En-Passant depend on more than just the current board position: they require the knowledge of a history of moves. To implement these, FEN notation might be a logical way to approach move order history.

Menus and other game features must also be implemented in order to complete the project.