Added piece drag and drop, and updated game stored board position.
Bug exists when moving piece to square a piece was previously on: piece goes to the squaare teh first piece moved to as well. EX. if knight moves from b8 to c6, and rook moves from a8 to b8, the rook teleports to c6.

Piece does not actually stay on the other square, and teleports to it's true square right after it is clicked and unclicked. Stored board position does not acknowledge weird movement.